update and reorganise SpanOptions unit tests
